Making MSVC happy

+++ lib/inc/sys_string/impl/unicode/mappings.h
@@ -336,7 +336,7 @@ namespace sysstr::util::unicode
                 return 0;
     
             if (res > 0x0FFF)
-                return res >> 12;
+                return uint8_t(res >> 12);
     
             size_t value_offset = ((size_t(c) - res) & 0x0FFF) - 1;
             uint32_t value = values[value_offset];
